requirements.txt
onvif-zeep==0.2.12
zeep==4.2.1
python-dotenv==1.0.1
